refactor: consolidate permission endpoints into frappe/core/api/permissions.py
Move the permission manager page endpoints (get_roles_and_doctypes,
get_permissions, add, update, remove, reset, get_users_with_role,
get_standard_permissions, get_permission_logs), the User Permission
endpoints (get_user_permissions, get_applicable_for_doctype_list,
check_applicable_doc_perm, clear_user_permissions, add_user_permissions)
and role_query into frappe/core/api/permissions.py, marked
@frappe.public(group="Permissions") with full annotations and
docstrings. The over-generic rule endpoints got proper names
(add/update/remove -> add/update/remove_permission_rule, reset ->
reset_permissions).

get_user_permissions' client-side self-restriction previously matched
only the bare "get_user_permissions" cmd shorthand; it now covers all of
the endpoint's cmd paths, so HTTP callers can no longer read other
users' permission list via the dotted paths (server-side callers are
unaffected).

Old dotted paths keep working via lazy module __getattr__ aliases;
in-repo JS (including the permission manager page's module/page-style
frappe.call invocations, now on explicit canonical paths) and imports
were updated.
