+++ frappe/core/doctype/doctype/doctype.js
@@ -38,7 +38,7 @@ frappe.ui.form.on("DocType", {
 		frm.set_query("role", "permissions", function (doc) {
 			if (doc.custom && frappe.session.user != "Administrator") {
 				return {
-					query: "frappe.core.doctype.role.role.role_query",
+					query: "frappe.core.api.permissions.role_query",
 				};
 			}
 		});
